chore: rename waterfall filter files/2
